better accesibility and some refactoring
