+++ src/pages/api/holdings.ts
@@ -1,9 +1,12 @@
 import { NextApiRequest, NextApiResponse } from 'next'
 import getHoldings from "src/service/holdings"
+
 export default async function(req: NextApiRequest, res: NextApiResponse) {
   try {
     if (req.method === 'GET') {
+      const start = Date.now()
       const holdings = await getHoldings()
+      res.setHeader("Server-Timing", `ms;dur=${Date.now() - start}`)
       res.json(holdings)
     } else {
       res.status(405)
