+++ src/styles/bottomBorder.ts
@@ -1,0 +1,7 @@
+
+import { css } from '@emotion/core'
+export default css({
+  borderBottom: 1,
+  borderBottomColor: 'rgba(46, 51, 56, 0.3)',
+  borderBottomStyle: 'solid',
+})
