+++ src/utils/fetcher.tsx
@@ -1,0 +1,5 @@
+
+export async function fetcher(url: string) {
+  const response = await fetch(url)
+  return response.json()
+}
