+++ src/utils/skipZeros.tsx
@@ -1,0 +1,4 @@
+import { TokenModel } from "src/service/Data"
+
+
+export const skipZeros = (token: TokenModel) => !isNaN(token.units) && token.value != 0
